refactor(frontend): main.js を機能別モジュールに分割 (#136)
Phase 1 のタスク C。元 798 行の `frontend/src/main.js` を機能別 ESM
モジュールに分割する。**振る舞いは一切変えていない** (関数の中身・実行
順序・シグネチャすべて元のまま、export/import 追加と関数移動のみ)。

分割マッピング:
- `lib/state.js`         共有状態 (config / guiSettings / *_PRESETS) と DISPATCH_KEYS
- `lib/tauri.js`         `window.__TAURI__.*` ラッパ (Phase 2 の差し替え点)
- `lib/utils.js`         escapeHtml
- `lib/dispatch-key.js`  createDispatchKeySelect / validateDispatchKeys
- `lib/tabs.js`          タブ切替リスナ
- `lib/shortcuts.js`     キーボードショートカット
- `lib/config-io.js`     loadConfig / renderConfig / collectConfig + 適用/リセット/初期値
- `forms/timestamp.js`   タイムスタンプフォーム
- `forms/search.js`      検索フォーム + プリセットモーダル
- `forms/folders.js`     フォルダフォーム
- `forms/apps.js`        アプリ切替フォーム + プロセスピッカーモーダル
- `forms/general.js`     全般 (autostart / export-import / help / quit / kanata / updater)
- `main.js`              エントリーポイント (~50 行、init 呼び出しのみ)

state は `let` 直 export を避け getter/setter 経由で共有 (ESM の export
binding 制約に対する素朴な回避)。循環依存なし、Vite build (18 modules)
成功を確認済み。
