+++ muhenkan-switch/frontend/src/lib/utils.js
@@ -1,0 +1,4 @@
+// ── Utility ──
+export function escapeHtml(str) {
+  return str.replace(/&/g, "&amp;").replace(/</g, "&lt;").replace(/>/g, "&gt;").replace(/"/g, "&quot;");
+}
